# Replacing type-check-and-cast sequence with a faster approach