--- a/src/ARMeilleure/Instructions/InstEmitSimdMemory32.cs
+++ b/src/ARMeilleure/Instructions/InstEmitSimdMemory32.cs
@@ -53,10 +53,8 @@ namespace ARMeilleure.Instructions
 
         public static void EmitVStoreOrLoadN(ArmEmitterContext context, int count, bool load)
         {
-            if (context.CurrOp is OpCode32SimdMemSingle)
+            if (context.CurrOp is OpCode32SimdMemSingle op)
             {
-                OpCode32SimdMemSingle op = (OpCode32SimdMemSingle)context.CurrOp;
-
                 int eBytes = 1 << op.Size;
 
                 Operand n = context.Copy(GetIntA32(context, op.Rn));