--- a/src/Ryujinx.HLE/HOS/Services/Audio/AudioRenderer/AudioRenderer.cs
+++ b/src/Ryujinx.HLE/HOS/Services/Audio/AudioRenderer/AudioRenderer.cs
@@ -55,9 +55,9 @@ namespace Ryujinx.HLE.HOS.Services.Audio.AudioRenderer
 
             if (resultCode == ResultCode.Success)
             {
-                if (outEvent is AudioKernelEvent)
+                if (outEvent is AudioKernelEvent evt)
                 {
-                    systemEvent = ((AudioKernelEvent)outEvent).Event;
+                    systemEvent = evt.Event;
                 }
                 else
                 {