Because the `is` operator performs a cast if the object is not null, using `is` to check type and then casting the same argument to that type, necessarily performs two casts.

Benchmark:
```cs
[SimpleJob]
[MemoryDiagnoser]
[Orderer(SummaryOrderPolicy.FastestToSlowest)]
[RankColumn]
public class Test
{
    private readonly object obj = 123;

    [Benchmark]
    public void TypeCheckTwice()
    {
        if (obj is int)
        {
            int castedInt = (int)obj;
        }
    }

    [Benchmark]
    public void TypeCheckOnce()
    {
        if (obj is int castedInt)
        {
        }
    }
}
```

Results:
```sh
|         Method |      Mean |     Error |    StdDev |    Median | Rank | Allocated |
|--------------- |----------:|----------:|----------:|----------:|-----:|----------:|
|  TypeCheckOnce | 0.7835 ns | 0.0646 ns | 0.1865 ns | 0.7197 ns |    1 |         - |
| TypeCheckTwice | 0.8953 ns | 0.0768 ns | 0.2155 ns | 0.7979 ns |    2 |         - |
```