@TSRBerry 2023-06-28
I don't think the benchmark code is valid here, since you are missing the assignment in the second test. This might lead to incorrect and misleading results.